requirements.txt
python-telegram-bot==20.3
pdfplumber==0.9.0
gspread==6.7.1
google-auth==2.21.0
